requirements.txt
python-telegram-bot==20.7
gspread
oauth2client
google-api-python-client
google-auth
Pillow
python-dotenv